// client/src/services/api/types/response.ts
export interface Response<T> {
  data: T;
}

export interface Collection<T> {
  data: T[];
  links: {
    first: string;
    last: string;
    prev: string | null;
    next: string | null;
  };
  meta: {
    current_page: number;
    from: number;
    last_page: number;
    links: {
      url: string | null;
      label: string;
      active: boolean;
    }[];
    path: string;
    per_page: number;
    to: number;
    total: number;
  },
}

export interface CollectionRequest {
  page: number;
}